base/server: move store_spec from server to base
This move will allow us to configure testclusters with store level granularity
which would otherwise cause circular dependencies.
