+++ base/store_spec.go
@@ -14,7 +14,7 @@
 //
 // Author: Bram Gruneir (bram+code@cockroachlabs.com)
 
-package server
+package base
 
 import (
 	"bytes"
@@ -26,7 +26,6 @@ import (
 
 	"github.com/dustin/go-humanize"
 
-	"github.com/cockroachdb/cockroach/config"
 	"github.com/cockroachdb/cockroach/roachpb"
 	"github.com/cockroachdb/cockroach/util/humanizeutil"
 )
@@ -34,7 +33,11 @@ import (
 // This file implements method receivers for members of server.Context struct
 // -- 'Stores' and 'JoinList', which satisfies pflag's value interface
 
-var minimumStoreSize = 10 * config.DefaultZoneConfig().RangeMaxBytes
+// MinimumStoreSize is the smallest size in bytes that a store can have. This
+// number was originally based on config's defaultZoneConfig's RangeMaxBytes,
+// which is extremely stable and to avoid adding the dependency on config here,
+// it is just hard coded to 640MiB.
+const MinimumStoreSize = 10 * 64 << 20
 
 // StoreSpec contains the details that can be specified in the cli pertaining
 // to the --store flag.
@@ -166,9 +169,9 @@ func newStoreSpec(value string) (StoreSpec, error) {
 				if err != nil {
 					return StoreSpec{}, fmt.Errorf("could not parse store size (%s) %s", value, err)
 				}
-				if ss.SizeInBytes < minimumStoreSize {
+				if ss.SizeInBytes < MinimumStoreSize {
 					return StoreSpec{}, fmt.Errorf("store size (%s) must be larger than %s", value,
-						humanizeutil.IBytes(minimumStoreSize))
+						humanizeutil.IBytes(MinimumStoreSize))
 				}
 			}
 		case "attrs":
